+++ test/blackbox-tests/test-cases/tests-stanza/run.t
@@ -11,15 +11,15 @@
   Entering directory 'plural'
       ocamldep .expect_test.eobjs/expect_test.ml.d
       ocamldep .expect_test.eobjs/regular_test.ml.d
-        ocamlc .expect_test.eobjs/expect_test.{cmi,cmo,cmt}
-      ocamlopt .expect_test.eobjs/expect_test.{cmx,o}
-      ocamlopt expect_test.exe
-   expect_test expect_test.output
         ocamlc .expect_test.eobjs/regular_test.{cmi,cmo,cmt}
       ocamlopt .expect_test.eobjs/regular_test.{cmx,o}
       ocamlopt regular_test.exe
   regular_test alias runtest
   regular test
+        ocamlc .expect_test.eobjs/expect_test.{cmi,cmo,cmt}
+      ocamlopt .expect_test.eobjs/expect_test.{cmx,o}
+      ocamlopt expect_test.exe
+   expect_test expect_test.output
   $ dune runtest --diff-command false --root generated --display short
   Entering directory 'generated'
       ocamldep .generated.eobjs/generated.ml.d